test: migrate tests from full to light and update coverage
- Move W8A8 dynamic EPLB test from full/two_cards to light/two_card
- Move 310P tests (dense, moe, vl, embedding, scoring, classification)
  from full to light with proper directory structure
- Move long_sequence tests (accuracy, basic, mtp, chunked_prefill,
  prefix_caching) from full to light/four_card
- Move deepseek_v4 test from full to light/four_card
- Move FA3, camem, cpu_offloading, xlite tests from full to light
- Add __init__.py for 310p and lora subdirectories
- Update coverage.md to reflect all 77 light tests with feature matrix
