+++ tests/e2e/pull_request/full/two_cards/test_qwen3_moe.py
@@ -16,15 +16,12 @@
 # This file is a part of the vllm-ascend project.
 # Adapted from vllm/tests/basic_correctness/test_basic_correctness.py
 #
-import json
 import os
 from unittest.mock import patch
 
-import openai
 import pytest
-from vllm.utils.network_utils import get_open_port
 
-from tests.e2e.conftest import RemoteOpenAIServer, VllmRunner
+from tests.e2e.conftest import VllmRunner
 from vllm_ascend.utils import vllm_version_is
 
 
@@ -101,52 +98,3 @@ def test_qwen3_moe_distributed_tp2_ep2_mrv2(
 
     for i in range(len(vllm_output)):
         assert golden_results[i] == vllm_output[i][1]
-
-
-@pytest.mark.asyncio
-async def test_qwen3_moe_w8a8_distributed_tp2_ep_dynamic_eplb():
-    model = "vllm-ascend/Qwen3-30B-A3B-W8A8"
-    port = get_open_port()
-    compilation_config = json.dumps({"cudagraph_capture_sizes": [8]})
-    server_args = [
-        "--max_model_len",
-        "8192",
-        "--tensor_parallel_size",
-        "2",
-        "--enable_expert_parallel",
-        "--quantization",
-        "ascend",
-        "--port",
-        str(port),
-        "--compilation-config",
-        compilation_config,
-    ]
-    env_dict = {"HCCL_BUFFSIZE": "1024"}
-    with RemoteOpenAIServer(model, server_args, server_port=port, auto_port=False, env_dict=env_dict) as server:
-        client = server.get_async_client()
-        batch = await client.completions.create(
-            model=model, prompt="What is deeplearning?", max_tokens=400, temperature=0, top_p=1.0, n=1
-        )
-        gt_choices: list[openai.types.CompletionChoice] = batch.choices
-
-    # dynamic eplb test
-    # Since pytest runs as a daemon, it conflicts with the dynamic eplb manager
-    # during initialization in offline mode, so the online mode is used instead.
-    env_dict.update({"DYNAMIC_EPLB": "true"})
-    additional_config = {
-        "eplb_config": {
-            "dynamic_eplb": True,
-            "expert_heat_collection_interval": 100,
-            "algorithm_execution_interval": 20,
-            "num_redundant_experts": 2,
-            "eplb_policy_type": 2,
-        }
-    }
-    server_args.extend(["--additional-config", json.dumps(additional_config)])
-    with RemoteOpenAIServer(model, server_args, server_port=port, auto_port=False, env_dict=env_dict) as server:
-        client = server.get_async_client()
-        batch = await client.completions.create(
-            model=model, prompt="What is deeplearning?", max_tokens=400, temperature=0, top_p=1.0, n=1
-        )
-        eplb_choices: list[openai.types.CompletionChoice] = batch.choices
-    assert gt_choices[0].text == eplb_choices[0].text, f"{gt_choices[0].text=} \n {eplb_choices[0].text=}"
