+++ tests/e2e/pull_request/light/one_card/_310p/test_embedding_310p.py
